I Made Cointoss
Cointoss has been programmed in a very general way where it asks for heads or tails and then randomly chooses heads or tails and if your guess matches what the program chose then you win the bet money. If not you lose the bet money.

How it is implemented into the other programs is a mystery to me so go ahead and look at it and see if there is anything that needs to be changed in order to be used in the program as a whole.